Revert "fix(product): path references unit tests"
This reverts commit 44ede4bde15079d09ee906126008fb9cb9f18288.
