+++ package.json
@@ -72,7 +72,7 @@
     "@types/helmet": "^0.0.48",
     "@types/jest": "^29.5.14",
     "@types/jsonwebtoken": "^9.0.9",
-    "@types/node": "^20.19.8",
+    "@types/node": "^20.3.1",
     "@types/supertest": "^2.0.12",
     "@types/winston": "^2.4.4",
     "@typescript-eslint/eslint-plugin": "^8.28.0",
@@ -89,9 +89,9 @@
     "prisma-erd-generator": "^2.0.4",
     "source-map-support": "^0.5.21",
     "supertest": "^6.3.3",
-    "ts-jest": "^29.4.0",
+    "ts-jest": "^29.1.0",
     "ts-loader": "^9.4.3",
-    "ts-node": "^10.9.2",
+    "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.2.0",
     "typescript": "^5.1.3"
   }
